requirements.txt
pinecone-client
tqdm
python-dotenv
requests
transformers
torch
